Update using latest BestieTemplate
