+++ RemoteBMI.jl/.editorconfig
@@ -5,6 +5,14 @@ root = true
 end_of_line = lf
 insert_final_newline = true
 charset = utf-8
-indent_size = 2
 indent_style = space
 trim_trailing_whitespace = true
+
+[*.jl]
+indent_size = 4
+
+[*.md]
+indent_size = 2
+
+[*.{yml,toml,json}]
+indent_size = 2
