+++ RemoteBMI.jl/.github/workflows/Docs.yml
@@ -35,19 +35,23 @@ jobs:
           version: "1"
       - name: Use Julia cache
         uses: julia-actions/cache@v2
-      - run: |
+      - name: Instantiate environment with development version of the package
+        run: |
           julia --project=docs -e '
             using Pkg
             Pkg.develop(PackageSpec(path=pwd()))
             Pkg.instantiate()'
-      - run: |
+      - name: Run doctest
+        run: |
           julia --project=docs -e '
             using Documenter: DocMeta, doctest
             using RemoteBMI
             DocMeta.setdocmeta!(RemoteBMI, :DocTestSetup, :(using RemoteBMI); recursive=true)
             doctest(RemoteBMI)'
-      - run: julia --project=docs docs/make.jl
+      - name: Generate and deploy documentation
+        run: julia --project=docs docs/make.jl
         env:
           JULIA_PKG_SERVER: ""
           GITHUB_TOKEN: ${{ secrets.GITHUB_TOKEN }}
           DOCUMENTER_KEY: ${{ secrets.DOCUMENTER_KEY }}
+          GKSwstype: "100" # https://discourse.julialang.org/t/generation-of-documentation-fails-qt-qpa-xcb-could-not-connect-to-display/60988
