+++ RemoteBMI.jl/.github/workflows/Test.yml
@@ -5,7 +5,6 @@ on:
     branches:
       - main
     tags: ["*"]
-  
   workflow_dispatch:
 
 jobs:
@@ -23,13 +22,10 @@ jobs:
       fail-fast: false
       matrix:
         version:
-          - "1.6"
+          - "lts"
           - "1"
         os:
           - ubuntu-latest
-          #- macOS-latest
-          #- windows-latest
-        arch:
+          arch:
           - x64
         allow_failure: [false]
-
