+++ RemoteBMI.jl/.gitignore
@@ -1,7 +1,12 @@
 *.jl.*.cov
 *.jl.cov
 *.jl.mem
-Manifest.toml
-docs/build/
 *.rej
+.DS_Store
+.benchmarkci
+Manifest.toml
+benchmark/*.json
+coverage
+docs/build/
+env
 node_modules
