+++ RemoteBMI.jl/CODE_OF_CONDUCT.md
@@ -60,7 +60,7 @@ representative at an online or offline event.
 ## Enforcement
 
 Instances of abusive, harassing, or otherwise unacceptable behavior may be
-reported to the community leaders responsible for enforcement at
+reported to the community leaders responsible for enforcement:
 `s.verhoeven@esciencecenter.nl`.
 All complaints will be reviewed and investigated promptly and fairly.
