+++ RemoteBMI.jl/Project.toml
@@ -12,4 +12,4 @@ TimeZones = "f269a46b-ccf7-5d73-abea-4c690281aa53"
 URIs = "5c2747f8-b7ea-4ff2-ba2e-563bfd36b1d4"
 
 [compat]
-julia = "1.6"
+julia = "1.10"
