+++ RemoteBMI.jl/docs/src/91-developer.md
@@ -86,7 +86,6 @@ We try to keep a linear history in this repo, so it is important to keep your br
     Try to create "atomic git commits" (recommended reading: [The Utopic Git History](https://blog.esciencecenter.nl/the-utopic-git-history-d44b81c09593)).
 
 - Make sure the tests pass.
-
 - Fetch any `main` updates from upstream and rebase your branch, if necessary:
 
   ```bash
