+++ RemoteBMI.jl/docs/src/index.md
@@ -6,4 +6,3 @@ CurrentModule = RemoteBMI
 
 Documentation for [RemoteBMI](https://github.com/eWaterCycle/RemoteBMI.jl).
 
-
